로그아웃 오류 수정

See merge request s11-webmobile1-sub2/S11P12A706!77

+++ battle-people-server/src/main/java/com/woowahanrabbits/battle_people/domain/user/controller/AuthController.java
@@ -53,8 +53,11 @@ public class AuthController {
 	}
 
 	@DeleteMapping("/logout")
-	public ResponseEntity<?> logout(@CookieValue("access") String access, HttpServletResponse response) {
-		userTokenRepository.deleteByUserId(jwtUtil.extractUserId(access));
+	public ResponseEntity<?> logout(@CookieValue(value = "access", required = false) String access,
+		HttpServletResponse response) {
+		if (access != null) {
+			userTokenRepository.deleteByUserId(jwtUtil.extractUserId(access));
+		}
 
 		HttpUtils.deleteCookies(
 			response,
